moved integration into index+separated plugin code
